Move PrincipalFactor and UserFactory to the api module
