+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/authentication/UserFactory.java
@@ -4,9 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.internal.subject;
-
-import io.kroxylicious.proxy.authentication.User;
+package io.kroxylicious.proxy.authentication;
 
 public class UserFactory implements PrincipalFactory<User> {
     @Override
