+++ kroxylicious-api/src/main/resources/META-INF/services/io.kroxylicious.proxy.authentication.PrincipalFactory
@@ -4,4 +4,4 @@
 # Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
 #
 
-io.kroxylicious.proxy.internal.subject.UserFactory
+io.kroxylicious.proxy.authentication.UserFactory
