+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/subject/PrincipalAdder.java
@@ -12,6 +12,7 @@ import java.util.function.Function;
 import java.util.stream.Stream;
 
 import io.kroxylicious.proxy.authentication.Principal;
+import io.kroxylicious.proxy.authentication.PrincipalFactory;
 
 public record PrincipalAdder(
                              Function<Object, Stream<String>> extractor,
